Depend on rustc-serialize from crates.io since Rust's in-tree serialize crate has been deprecated
This also fixes a problem where depending on serialize would pull
in the also deprecated log and regex crates from the Rust distribution
instead of their Crates.io versions. This could lead to dependency
issues for projects using Rust-Crypto that needed to use the
Crates.io versions of those libraries.