Clean up the Salsa20/XSalsa20 code to better reflect the papers.
Storing the state as an array of 'u32's is cleaner and minutely faster,
though the performance difference is negligible.  The main gains here
are readability related.

To further validate that nothing was broken, one of the NaCl test
vectors was also imported (256 bit, 64 bit nonce, examines 4 MiB of
keystream).