Remove use of SIMD types, for now
The simd module should be able to support using real SIMD types
when compiled with a nightly and 'fake' SIMD types when compiled
against stable. However, I'm hitting some ICE that I don't know
how to workaround when trying to use the real SIMD types with this
approach. So, disable SIMD types until we can figure that out.